🔨fix: useModal의 button1 clickHandler 달 수 있도록 수정
선택적으로 clickHandler 속성 선택 가능
