+++ grass-diary/src/state/modal/useModal.ts
@@ -1,21 +1,6 @@
 import { useModalButtonActions } from './ModalButtonStore';
 import { useModalActions } from './ModalStore';
 
-type Setting = {
-  title: string;
-  content: string;
-};
-
-type Button2 = {
-  active: boolean;
-  text: string;
-  color?: string;
-  interaction?: string;
-  clickHandler: () => void;
-};
-
-type Button1 = Pick<Button2, 'active' | 'text' | 'color' | 'interaction'>;
-
 export const useModal = () => {
   const { setLogin, setActive, setSetting } = useModalActions();
   const { setResetActive, setButton1, setButton2 } = useModalButtonActions();
